+++ src/targeted/targets/action-type-target.mjs
@@ -1,7 +1,7 @@
 import { MODULE_NAME } from '../../consts.mjs';
 import { radioInput } from '../../handlebars-handlers/bonus-inputs/radio-input.mjs';
 import { traitInput } from '../../handlebars-handlers/trait-input.mjs';
-import { isMelee, isNatural, isNaturalSecondary, isPhysical, isRanged, isSpell, isThrown, isWeapon } from '../../util/action-type-helpers.mjs';
+import { isCmb, isMelee, isNatural, isNaturalSecondary, isPhysical, isRanged, isSpell, isThrown, isWeapon } from '../../util/action-type-helpers.mjs';
 import { listFormat } from '../../util/list-format.mjs';
 import { localize, localizeBonusLabel } from '../../util/localize.mjs';
 import { toArray } from '../../util/to-array.mjs';
@@ -11,6 +11,7 @@ import { BaseTarget } from './_base-target.mjs';
 /** @typedef {keyof typeof filterTypes} FilterType */
 
 const filterTypes = /** @type {const} */ ({
+    ['is-cmb']: { label: '', filter: isCmb },
     ['is-melee']: { label: '', filter: isMelee },
     ['is-natural']: { label: '', filter: isNatural },
     ['is-natural-secondary']: { label: '', filter: isNaturalSecondary },
@@ -145,12 +146,12 @@ export class ActionTypeTarget extends BaseTarget {
      * @param {ItemPF} options.item
      */
     static showInputOnItemSheet({ html, isEditable, item }) {
-        /** @type {{[key: string]: string}} */
-        const choices = {};
-        Object.keys(filterTypes).forEach((f) => {
+        const choices = Object.keys(filterTypes).map((f) => {
             const key = /** @type {FilterType} */ (f);
-            choices[key] = filterTypes[key].label;
-        });
+            return { id: key, label: filterTypes[key].label };
+        })
+            .sort((a, b) => a.label.localeCompare(b.label))
+            .reduce((obj, { id, label }) => ({ ...obj, [id]: label }), {});
 
         const radioValues = [
             { id: all, label: localize(`target-toggle.${all}`) },