+++ types/my-types.d.ts
@@ -160,6 +160,7 @@ import { SpellSubschoolTarget } from '../src/targeted/targets/spell-subschool-ta
 import { WeaponGroupTarget } from '../src/targeted/targets/weapon-group-target.mjs';
 import { WeaponTypeTarget } from '../src/targeted/targets/weapon-type-target.mjs';
 import {
+    isCmb,
     isMelee,
     isNatural,
     isNaturalSecondary,
@@ -477,6 +478,7 @@ declare global {
         /** various utility helper methods and classes used throughout the mod */
         utils: {
             actionTypeHelpers: {
+                isCmb: typeof isCmb;
                 isMelee: typeof isMelee;
                 isNatural: typeof isNatural;
                 isNaturalSecondary: typeof isNaturalSecondary;